@@ -1,9 +1,9 @@
-import React, { useState } from "react";
+import React, { useState, useEffect } from "react";
 import { PlayerType, Poker } from "../Logic/types";
 import Hand from "./Hand";
-import { playerBet } from "../Logic/gameplay";
 import styled from "@emotion/styled";
-import { raise, fold, nextPlayer } from "../Logic/playerActions";
+import { raise, fold, call, nextPlayer } from "../Logic/actions";
+import { cloneDeep } from "lodash";
 
 interface Props {
   player: PlayerType;
@@ -23,6 +23,14 @@ let Actions = styled.div`
   align-items: center;
 `;
 
+let BetInput = styled.div`
+  display: flex;
+  flex-direction: row;
+  height: 40px;
+  align-items: center;
+  justify-content: center;
+`;
+
 function Player({
   player,
   dealer,
@@ -35,10 +43,19 @@ function Player({
   const [bet, setBet] = useState(20);
   const [error, setError] = useState(null);
 
+  useEffect(() => {
+    if (!player.isActive && index === game.actionOn) {
+      const newGame = cloneDeep(game);
+      newGame.actionOn = nextPlayer(newGame, index);
+      saveGame(newGame);
+    }
+  }, [game]);
+
   const sendAction = (action: any) => {
     const { result, error } = action(game, index, bet);
     if (result) {
-      result.actionOn = nextPlayer(game, index);
+      result.actionOn = nextPlayer(result, index);
+      console.log("result from action: ", action, result);
       saveGame(result);
     } else {
       setError(error);
@@ -48,19 +65,45 @@ function Player({
   return (
     <div key={index}>
       <h2>{player.name}</h2>
-      <h3>{index === dealer ? "dealer" : "player"}</h3>
-      <h3>Bet: </h3>
+      <h3>
+        {index === dealer
+          ? "Dealer"
+          : index === dealer + 1 || dealer === game.players.length - 1
+          ? "Small blind"
+          : index === dealer + 2 || dealer === game.players.length - 2
+          ? "Big blind"
+          : " "}
+      </h3>
       <h2>{player.stack}</h2>
       <Hand hand={player.cards} active={player.isActive} />
       {winner.includes(index) && <h3>Winner!</h3>}
       <Actions>
+        <BetInput>
+          <h3>Bet: </h3>
+          <input
+            type="number"
+            min={game.currentBet}
+            max={player.stack}
+            //class="slider"
+            onChange={(event: any) =>
+              setBet(Number.parseInt(event.target.value))
+            }
+            value={bet}
+            disabled={game.actionOn !== index}
+          />
+        </BetInput>
         <button
           onClick={() => sendAction(raise)}
           disabled={game.actionOn !== index}
         >
-          Raise
+          Raise: {bet}
         </button>
-        <button disabled={game.actionOn !== index}>Check/Call </button>
+        <button
+          onClick={() => sendAction(call)}
+          disabled={game.actionOn !== index}
+        >
+          Check/Call{" "}
+        </button>
         <button
           onClick={() => sendAction(fold)}
           disabled={game.actionOn !== index}
